move PipelineRetryStrategy to pipeline.rs
