+++ glide-core/redis-rs/redis-test/src/lib.rs
@@ -26,8 +26,9 @@
 use std::collections::VecDeque;
 use std::sync::{Arc, Mutex};
 
-use redis::cluster_async::PipelineRetryStrategy;
-use redis::{Cmd, ConnectionLike, ErrorKind, Pipeline, RedisError, RedisResult, Value};
+use redis::{
+    Cmd, ConnectionLike, ErrorKind, Pipeline, PipelineRetryStrategy, RedisError, RedisResult, Value,
+};
 
 #[cfg(feature = "aio")]
 use futures::{future, FutureExt};
